refactor(kyria): Refactor paren/brace behaviors and combos
There was originally a bug that caused the behavior I originally
intended to not work as expected, which meant that the behavior and
combo names wasn't reflecting the actual behavior. Now that the bug has
been fixed it turns out I prefer the "accidental" behavior over what I
originally intended.